+++ CHANGELOG.md
@@ -11,6 +11,11 @@ Prefix your message with one of the following:
 - [Security] in case of vulnerabilities.
 -->
 
+## Unreleased
+
+- [Changed] Add newline to multiline strings if they don't end with one.
+- [Fixed] Ensure multiline strings are passed as the argument.
+
 ## v0.0.4
 
 - [Fixed] Fix how Thor handle defaults for arrays.